refactor: 별점 핸들링 컨텍스트로 리팩터링
