--- frontend/src/hooks/review/useStarRating.ts
@@ -1,22 +1,0 @@
-import { useState } from 'react';
-
-const useStarRating = () => {
-  const [rating, setRating] = useState(0);
-  const [hovering, setHovering] = useState(0);
-
-  const handleRating = (starIndex: number) => {
-    setRating(starIndex);
-  };
-
-  const handleMouseEnter = (starIndex: number) => {
-    setHovering(starIndex);
-  };
-
-  const handleMouseLeave = () => {
-    setHovering(0);
-  };
-
-  return { rating, setRating, hovering, handleRating, handleMouseEnter, handleMouseLeave };
-};
-
-export default useStarRating;
