+++ frontend/src/hooks/review/useStarRatingHover.ts
@@ -1,0 +1,17 @@
+import { useState } from 'react';
+
+const useStarRatingHover = () => {
+  const [hovering, setHovering] = useState(0);
+
+  const handleMouseEnter = (starIndex: number) => {
+    setHovering(starIndex);
+  };
+
+  const handleMouseLeave = () => {
+    setHovering(0);
+  };
+
+  return { hovering, handleMouseEnter, handleMouseLeave };
+};
+
+export default useStarRatingHover;
